requirements.txt
flask==3.1.1
flask-cors==6.0.1
openai-whisper==20240930
selenium==4.33.0
pysrt==1.1.2
google-cloud-translate==3.20.3
requests==2.32.4
beautifulsoup4==4.12.3

